\name{getNumericInput}
\alias{getNumericInput}
\title{Get Numeric Input}
\usage{
  getNumericInput(lower = -Inf, upper = Inf)
}
\arguments{
  \item{lower}{Numeric. The lower endpoint of acceptable
  values.}

  \item{upper}{Numeric. The upper endpoint of acceptable
  values.}
}
\description{
  Provide a prompt that gets numeric input in a certain
  range. -Inf and Inf are acceptable endpoints of the range
  to tell the function that you don't actually have an
  upper or lower end point.
}

