\name{hangman}
\alias{hangman}
\title{Hangman Game}
\usage{
  hangman(reset.score = FALSE)
}
\arguments{
  \item{reset.score}{logical.  If \code{TRUE} resets the
  scoring for the R session.}
}
\value{
  Returns a vector of character strings with escaped
  characters removed.
}
\description{
  The classic wordgame designed to waste hours of your
  time.
}
\author{
  Tyler Rinker
}

